reduces syscalls by changing to coarsetime when getting arrival times for pending withdraws

+++ src/transaction.rs
@@ -17,6 +17,6 @@ pub enum Transaction {
 }
 
 pub struct PendingWithdraw {
-    pub arrival_time: u128,
+    pub arrival_time: u64,
     pub tx: Tx,
 }
